fixes and updates
- cleaned up backend-job to move core components to internal directory
- fixes rate limiter to be by client IP and added additional logging
- upgraded node_modules in webui to latest
