+++ .gitignore
@@ -22,3 +22,5 @@ local.md
 
 *.bluge
 webui.fluentui
+
+ips.changes.*
