+++ webui/package.json
@@ -5,23 +5,23 @@
   "dependencies": {
     "@rsuite/icons": "^1.0.2",
     "@rsuite/responsive-nav": "^5.0.1",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.2.0",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.27.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.2.2",
     "customize-cra": "^1.0.0",
-    "less": "^4.1.2",
-    "less-loader": "^7.3.0",
+    "less": "^4.1.3",
+    "less-loader": "7.3.0",
     "react": "^18.2.0",
     "react-app-rewired": "^2.2.1",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.3.1",
-    "react-markdown": "^8.0.3",
-    "react-player": "^2.10.1",
+    "react-icons": "^4.7.1",
+    "react-markdown": "^8.0.4",
+    "react-player": "^2.11.0",
     "react-scripts": "^5.0.1",
     "remark-gfm": "^3.0.1",
-    "rsuite": "^5.16.1",
-    "web-vitals": "^2.1.4"
+    "rsuite": "^5.24.1",
+    "web-vitals": "^3.1.1"
   },
   "scripts": {
     "start": "react-app-rewired start",
